+++ src/http/index.ts
@@ -4,6 +4,7 @@ export * from "./createHttpApplication";
 export * from "./database/RequestLog";
 export * from "./error-codes";
 export * from "./events";
+export * from "./health";
 export * from "./request-controller";
 export * from "./uploaded-file";
 export * from "./uploads-config";